+++ configs/bootstrappers.php
@@ -1,10 +1,10 @@
 <?php
+use Opulence\Framework\Bootstrappers\Cryptography\CryptographyBootstrapper;
+use Opulence\Framework\Bootstrappers\Php\PhpBootstrapper;
 use Project\Bootstrappers\Databases\RedisBootstrapper;
 use Project\Bootstrappers\Databases\SqlBootstrapper;
 use Project\Bootstrappers\Events\DispatcherBootstrapper;
 use Project\Bootstrappers\Orm\UnitOfWorkBootstrapper;
-use Opulence\Framework\Bootstrappers\Cryptography\CryptographyBootstrapper;
-use Opulence\Framework\Bootstrappers\Php\PhpBootstrapper;
 
 /**
  * ----------------------------------------------------------