+++ configs/console/exceptions.php
@@ -1,0 +1,9 @@
+<?php
+use Opulence\Framework\Exceptions\Console\ExceptionRenderer;
+
+/**
+ * ----------------------------------------------------------
+ * Set the console exception renderer
+ * ----------------------------------------------------------
+ */
+return new ExceptionRenderer();