+++ configs/http/exceptions.php
@@ -1,0 +1,9 @@
+<?php
+use Opulence\Framework\Exceptions\Http\ExceptionRenderer;
+
+/**
+ * ----------------------------------------------------------
+ * Set the HTTP exception renderer
+ * ----------------------------------------------------------
+ */
+return new ExceptionRenderer();