+++ configs/logging.php
@@ -4,8 +4,11 @@ use Monolog\Logger;
 
 /**
  * ----------------------------------------------------------
- * Create a Monolog logger
+ * Create a PSR-3 logger
  * ----------------------------------------------------------
+ *
+ * Note: You may use any PSR-3 logger you'd like
+ * For convenience, the Monolog library is included here
  */
 $logger = new Logger("application");
 $logger->pushHandler(new ErrorLogHandler());