+++ resources/views/errors/400.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/Error") %>
+
+<% part("errorTitle") %>
+    Bad request
+<% endpart %>
+<% part("errorDescription") %>
+    That request you made was invalid.
+<% endpart %>