+++ resources/views/errors/401.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/Error") %>
+
+<% part("errorTitle") %>
+    Unauthorized
+<% endpart %>
+<% part("errorDescription") %>
+    You do not have permission to view this page.
+<% endpart %>