+++ resources/views/errors/500.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/Error") %>
+
+<% part("errorTitle") %>
+    Something broke
+<% endpart %>
+<% part("errorDescription") %>
+    Sorry about that.
+<% endpart %>