+++ resources/views/errors/Error.fortune.php
@@ -1,0 +1,8 @@
+<% extends("Master") %>
+
+<% part("content") %>
+    <h2><% show("errorTitle") %></h2>
+    <% part("errorDescription") %>
+        Something went wrong.
+    <% show %>
+<% endpart %>