+++ tests/src/Project/Http/ApplicationTestCase.php
@@ -1,8 +1,9 @@
 <?php
 namespace Project\Http;
 
-use Opulence\Applications\Application;
 use Opulence\Bootstrappers\ApplicationBinder;
+use Opulence\Exceptions\ExceptionHandler;
+use Opulence\Framework\Exceptions\Http\IHttpExceptionRenderer;
 use Opulence\Framework\Testing\PhpUnit\Http\ApplicationTestCase as BaseTestCase;
 use Opulence\Ioc\IContainer;
 
@@ -11,6 +12,40 @@ use Opulence\Ioc\IContainer;
  */
 class ApplicationTestCase extends BaseTestCase
 {
+    /** @var ExceptionHandler The exception handler used by HTTP applications */
+    private $exceptionHandler = null;
+    /** @var IHttpExceptionRenderer The exception renderer used by HTTP applications */
+    private $exceptionRenderer = null;
+
+    /**
+     * @inheritdoc
+     */
+    public function setUp()
+    {
+        /** @var $exceptionRenderer IHttpExceptionRenderer */
+        $exceptionRenderer = require __DIR__ . "/../../../../configs/http/exceptions.php";
+        $this->exceptionHandler = require __DIR__ . "/../../../../configs/exceptions.php";
+        $this->exceptionRenderer = $exceptionRenderer;
+
+        parent::setUp();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getExceptionHandler()
+    {
+        return $this->exceptionHandler;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function getExceptionRenderer()
+    {
+        return $this->exceptionRenderer;
+    }
+
     /**
      * @inheritdoc
      */
@@ -22,19 +57,10 @@ class ApplicationTestCase extends BaseTestCase
     /**
      * @inheritdoc
      */
-    protected function getKernelLogger()
-    {
-        return require __DIR__ . "/../../../../configs/http/logging.php";
-    }
-
-    /**
-     * @inheritdoc
-     */
     protected function setApplicationAndIocContainer()
     {
-        /** @var Application $application */
-        require __DIR__ . "/../../../../bootstrap/start.php";
-        $this->application = $application;
+        require __DIR__ . "/../../../../configs/paths.php";
+        $this->application = require __DIR__ . "/../../../../configs/application.php";
         /** @var IContainer $container */
         $this->container = $container;